+++ src/app/services/xmat-test-mocks-list.service.ts
@@ -41,6 +41,10 @@ export class XmatTestMocksListService extends XmatMocksListService {
         {
             method: this._methodsKeys.GET,
             url: this._servicesUrls.getRequestsByUid
+        },
+        {
+            method: this._methodsKeys.PATCH,
+            url: this._servicesUrls.testPatch
         }
     ];