+++ src/app/views/xmat-rest-examples/xmat-rest-examples.component.ts
@@ -1,23 +1,32 @@
-import {Component, OnDestroy, OnInit} from "@angular/core";
-import {XmatTestRestService} from "../../services/xmat-test-rest.service";
+import {Component, OnInit} from "@angular/core";
 import {HttpErrorResponse} from "@angular/common/http";
 import {XmatHttpParams, XmatHttpConfig} from "ngx-mat-lib";
-import {of} from "rxjs";
-import {delay} from "rxjs/operators";
+//
+import {XmatTestRestService} from "../../services/xmat-test-rest.service";
+
+export enum XmatRestStates {
+    RUNNING = "RUNNING",
+    IDLE = "IDLE",
+    SUCCESS = "SUCCESS",
+    FAIL = "FAIL"
+}
 
 @Component({
     selector: "xmat-test",
     templateUrl: "./xmat-rest-examples.component.html"
 })
 
-export class XmatRestExamplesComponent implements OnInit, OnDestroy {
+export class XmatRestExamplesComponent implements OnInit {
 
+    patchData: any;
+    states: typeof XmatRestStates = XmatRestStates;
+    //
     preDelay: number = 2000;
     postDelay: number = 3000;
     //
-    statusA: string = "idle";
-    statusB: string = "idle";
-    statusC: string = "idle";
+    statusA: XmatRestStates = XmatRestStates.IDLE;
+    statusB: XmatRestStates = XmatRestStates.IDLE;
+    statusC: XmatRestStates = XmatRestStates.IDLE;
     //
     dataA: any;
     dataB: any;
@@ -34,15 +43,35 @@ export class XmatRestExamplesComponent implements OnInit, OnDestroy {
 
     }
 
-    ngOnInit(): void {
-        this._callA();
-        // this._callB();
-        // this._callC();
-        // this._testQueueObj();
+    doPatch(): void {
+        this.patchData = void 0;
+        const config = this._rest.servicesConfigs.testPatch();
+        config.data = {
+            foo: "bar",
+            goo: "car"
+        };
+        this._rest.$http(config).subscribe(response => {
+            console.info("PATCH SUCCESS", response);
+            this.patchData = response;
+        }, (error: HttpErrorResponse) => {
+            console.info("PATCH FAIL", error);
+        });
     }
 
-    ngOnDestroy(): void {
+    doQueue(): void {
+        this._rest.$allMap({
+            first: this._configB,
+            second: this._configC
+        })
+        .subscribe((response: any) => {
+            console.info("$all response", response);
+        });
+    }
 
+    ngOnInit(): void {
+        this._callA();
+        this._callB();
+        this._callC();
     }
 
     reloadA(): void {
@@ -54,50 +83,40 @@ export class XmatRestExamplesComponent implements OnInit, OnDestroy {
     protected _callA(): void {
         const myConfigA = this._rest.servicesConfigs.postUsers();
         myConfigA.params = new XmatHttpParams(this.preDelay, this.postDelay);
-        this.statusA = "processing...";
+        this.statusA = XmatRestStates.RUNNING;
         this.dataA = this.timeA = void 0;
         const startA = +new Date();
         this._rest.$http(myConfigA).subscribe(response => {
-            console.info("GOT RESPONSE A", response);
+            console.info("GOT RESPONSE FOR callA", response);
             this.dataA = response;
-            this.statusA = "success";
+            this.statusA = XmatRestStates.SUCCESS;
             this.timeA = +new Date() - startA;
         }, () => {
-            this.statusA = "fail";
+            this.statusA = XmatRestStates.FAIL;
         });
     }
 
     protected _callB(): void {
-        this.statusB = "processing...";
+        this.statusB = XmatRestStates.RUNNING;
         this.dataB = void 0;
         this._rest.$http(this._configB).subscribe(response => {
             this.dataB = response;
-            this.statusB = "success";
+            this.statusB = XmatRestStates.SUCCESS;
         }, (error: HttpErrorResponse) => {
-            this.statusB = `fail ${error.status}`;
+            console.warn("callB errored", error);
+            this.statusB = XmatRestStates.FAIL;
         });
     }
 
     protected _callC(): void {
-        this.statusC = "processing...";
+        this.statusC = XmatRestStates.RUNNING;
         this.dataC = void 0;
         this._rest.$http(this._configC).subscribe(response => {
             this.dataC = response;
-            this.statusC = "success";
+            this.statusC = XmatRestStates.SUCCESS;
         }, (error: HttpErrorResponse) => {
-            this.statusC = `fail ${error.status}`;
+            console.warn("callC errored", error);
+            this.statusC = XmatRestStates.FAIL;
         });
     }
-
-    protected _testQueueObj(): void {
-
-        this._rest.$allMap({
-            first: this._configB,
-            second: this._configC
-        })
-        .subscribe((response: any) => {
-            console.info("$all response", response);
-        });
-    }
-
 }