+++ src/environments/environment.ts
@@ -1,5 +1,5 @@
 
 export const xmatEnvironment = {
-    appVersion: "7.2.9",
+    appVersion: "7.2.11",
     production: !1
 };